1107 - 가계부 단위 double -> long 형으로 변경/ 소소한 css 수정들 - 해진

+++ Semi/src/controller/plan/PlanViewController.java
@@ -99,7 +99,7 @@ public class PlanViewController extends HttpServlet {
 //		--------------------------------------------
 		
 		// 가계부 정보 가져오기
-		EnumMap<AccType, Double> accEnumMap = new EnumMap<AccType, Double>(AccType.class);
+		EnumMap<AccType, Long> accEnumMap = new EnumMap<AccType, Long>(AccType.class);
 	
 		// 카테고리 별
 		accEnumMap.put(AccType.airfare, pService.getAccountAccTpeCost(planParam, AccType.airfare));
@@ -114,7 +114,7 @@ public class PlanViewController extends HttpServlet {
 		req.setAttribute("accTypeCost", new Gson().toJson(accEnumMap));
 		
 		// 총합
-		double acc_total = pService.getAccountTotal(accEnumMap);
+		Long acc_total = pService.getAccountTotal(accEnumMap);
 		req.setAttribute("acc_total", acc_total);
 		
 		// view 폼 띄우기
